Move app higher order component to components dir
